src/maintenance_server/files/home/ansible/requirements.txt
paramiko
boto3
pyyaml
